put NSWindow enum constants into separate file
